个人中心的回收站 完善

+++ service/src/main/java/com/zone/service/Impl/ArticleServiceImpl.java
@@ -5,6 +5,7 @@ import com.github.pagehelper.Page;
 import com.github.pagehelper.PageHelper;
 import com.zone.constant.ArticleAboutConstant;
 import com.zone.constant.DeleteConstant;
+import com.zone.context.BaseContext;
 import com.zone.dto.ArticleEditDTO;
 import com.zone.dto.ArticlePublishDTO;
 import com.zone.dto.PageBean;
@@ -32,11 +33,10 @@ import java.util.List;
 
 @Service // 注解
 @Slf4j // 日志
-@RequiredArgsConstructor // 构造方法注解 注入bean对象
-@SuppressWarnings("ALL")
+@RequiredArgsConstructor // 构造方法注解
+@SuppressWarnings("ALL") // 忽略警告
 public class ArticleServiceImpl extends ServiceImpl<ArticleMapper, Article> implements ArticleService {
 
-    // 注入Mapper
     private final ArticleMapper articleMapper;
 
     private final ArticleCategoryMapper articleCategoryMapper;
@@ -62,8 +62,7 @@ public class ArticleServiceImpl extends ServiceImpl<ArticleMapper, Article> impl
                 .createTime(LocalDateTime.now()) // 创建时间
                 .updateTime(LocalDateTime.now()) // 更新时间
                 .deleteTime(ArticleAboutConstant.DEFAULT_ARTICLE_DELETE_TIME) // 删除时间
-                //TODO 根据JWT令牌解析Token 去获取当前创建人的ID 从而获取name 暂时写死
-                .creator(ArticleAboutConstant.DEFAULT_CREATOR) // 创建者
+                .creatorId(BaseContext.getCurrentId()) // 创建者ID
                 .status(ArticleAboutConstant.DEFAULT_STATUS) // 状态
                 .deleteStatus(DeleteConstant.ENABLE) // 删除状态
                 .build();
@@ -75,7 +74,6 @@ public class ArticleServiceImpl extends ServiceImpl<ArticleMapper, Article> impl
         // 2.添加文章类型article_category到数据库中
         ArticleCategory articleCategory = new ArticleCategory();
         articleCategory.setArticleId(article.getId());
-
         articlePublishDTO.getCategoryId().forEach(id -> {
             // 遍历添加
             articleCategory.setCategoryId(id);
@@ -83,11 +81,9 @@ public class ArticleServiceImpl extends ServiceImpl<ArticleMapper, Article> impl
         });
 
 
-
         // 3.添加文章标签article_tag到数据库中
         ArticleTag articleTag = new ArticleTag();
         articleTag.setArticleId(article.getId());
-
         articlePublishDTO.getTagId().forEach(id -> {
             // 遍历添加
             articleTag.setTagId(id);
@@ -112,39 +108,40 @@ public class ArticleServiceImpl extends ServiceImpl<ArticleMapper, Article> impl
                 .content(articleEditDTO.getContent()) // 内容
                 .articleViewPower(articleEditDTO.getArticleViewPower()) // 访问权限
                 .updateTime(LocalDateTime.now()) // 更新时间
+                .creatorId(BaseContext.getCurrentId()) // 创建者ID
                 .build();
         articleMapper.update(article);
 
-        // 2.根据文章ID删除文章分类
-        articleCategoryMapper.deleteByArticleId(articleEditDTO.getId());
-        // 3.根据文章ID删除标签分类
-        articleTagMapper.deleteByArticleId(articleEditDTO.getId());
-
-
-        // 3.添加文章分类
-        articleEditDTO.getCategoryId().forEach(id -> {
-            // 创建对象
-            ArticleCategory articleCategory = new ArticleCategory();
-            // 设置文章ID
-            articleCategory.setArticleId(articleEditDTO.getId());
-            // 设置分类ID
-            articleCategory.setCategoryId(id);
-            // 插入数据
-            articleCategoryMapper.insert(articleCategory);
-        });
-
-        // 4.添加标签分类
-        articleEditDTO.getTagId().forEach(id -> {
-            // 创建对象
-            ArticleTag articleTag = new ArticleTag();
-            // 设置文章ID
-            articleTag.setArticleId(articleEditDTO.getId());
-            // 设置标签ID
-            articleTag.setTagId(id);
-            // 插入数据
-            articleTagMapper.insert(articleTag);
-        });
-
+        log.info("articleEditDTO:{}", articleEditDTO);
+        // 2.根据文章ID删除/并添加新的 文章分类
+        if (articleEditDTO.getCategoryId() != null) {
+            articleCategoryMapper.deleteByArticleId(articleEditDTO.getId());
+            articleEditDTO.getCategoryId().forEach(id -> {
+                // 创建对象 articleCategory
+                ArticleCategory articleCategory = new ArticleCategory();
+                // 设置文章ID
+                articleCategory.setArticleId(articleEditDTO.getId());
+                // 设置分类ID
+                articleCategory.setCategoryId(id);
+                // 插入数据
+                articleCategoryMapper.insert(articleCategory);
+            });
+        }
+        // 3.根据文章ID删除/并添加新的 标签分类
+        if (articleEditDTO.getTagId() != null) {
+            articleTagMapper.deleteByArticleId(articleEditDTO.getId());
+            //TODO 不能使用循环来调用SQL语句
+            articleEditDTO.getTagId().forEach(id -> {
+                // 创建对象
+                ArticleTag articleTag = new ArticleTag();
+                // 设置文章ID
+                articleTag.setArticleId(articleEditDTO.getId());
+                // 设置标签ID
+                articleTag.setTagId(id);
+                // 插入数据
+                articleTagMapper.insert(articleTag);
+            });
+        }
         log.info("编辑文章成功");
     }
 
@@ -159,6 +156,7 @@ public class ArticleServiceImpl extends ServiceImpl<ArticleMapper, Article> impl
         Article article = Article.builder()
                 .deleteStatus(DeleteConstant.DISABLE)
                 .id(id)
+                .creatorId(BaseContext.getCurrentId())
                 .deleteTime(LocalDateTime.now())
                 .build();
 
@@ -173,19 +171,21 @@ public class ArticleServiceImpl extends ServiceImpl<ArticleMapper, Article> impl
      */
     @Override
     public PageResult search(PageSearchDTO pageSearchDTO) {
+        pageSearchDTO.setUserId(BaseContext.getCurrentId());
         log.info("根据文章题目进行模糊搜索:{}", pageSearchDTO);
 
         // 1.使用PageHelper分页查询
         PageHelper.startPage(pageSearchDTO.getPageNum(), pageSearchDTO.getPageSize());
+        // 获取文章时 只能获取个人文章 不能获取别人的文章
         Page<Article> articlePage = articleMapper.search(pageSearchDTO);
-        long total = articlePage.getTotal();
 
         // 2.根据文章ID获取文章分类
         List<ArticleVO> articleVOS = gainCategoryByArticleId(articlePage);
-
         log.info("根据文章题目进行模糊搜索成功");
 
+
         // 3.返回
+        long total = articlePage.getTotal();
         return new PageResult(total, articleVOS);
     }
 
@@ -196,13 +196,13 @@ public class ArticleServiceImpl extends ServiceImpl<ArticleMapper, Article> impl
     @Override
     public void deleteArticleByIds(List<Integer> ids) {
         // 1.删除文章
+        articleMapper.deleteBatchIds(ids);
 
-        ids.forEach(id->{
-            // 2.删除文章分类
-            articleCategoryMapper.deleteByArticleId(id);
-            // 3.删除文章标签
-            articleTagMapper.deleteByArticleId(id);
-        });
+        // 2.删除文章分类
+        articleCategoryMapper.deleteByArticleIds(ids);
+
+        // 3.删除文章标签
+        articleTagMapper.deleteByArticleIds(ids);
 
         log.info("批量删除文章成功");
     }
@@ -224,7 +224,7 @@ public class ArticleServiceImpl extends ServiceImpl<ArticleMapper, Article> impl
     public PageResult listArticle(PageBean pageBean) {
         // 获取所有人的文章
         // 1.进行分页查询
-        PageHelper.startPage(pageBean.getPageNum(),pageBean.getPageSize());
+        PageHelper.startPage(pageBean.getPageNum(), pageBean.getPageSize());
 
         // 2.获取文章
         Page<Article> articlePage = articleMapper.listArticle();
@@ -238,14 +238,14 @@ public class ArticleServiceImpl extends ServiceImpl<ArticleMapper, Article> impl
         });
 
         // 4.返回数据
-        return new PageResult(articlePage.getTotal(),articleManageVOS);
+        return new PageResult(articlePage.getTotal(), articleManageVOS);
     }
 
     /**
      * 根据文章ID获取文章分类
      */
     private List<ArticleVO> gainCategoryByArticleId(List<Article> articles) {
-
+        log.info("articles:{}", articles);
         List<ArticleVO> articleVOS = new ArrayList<>();
         articles.forEach(article -> {
             // 2.1创建对象
@@ -255,7 +255,7 @@ public class ArticleServiceImpl extends ServiceImpl<ArticleMapper, Article> impl
             // 2.2创建集合
             List<String> articleCategoryNameList = new ArrayList<>();
 
-            // 2.3根据查询文章ID 获取文章分类名称
+            // 2.3根据文章ID 获取文章分类名称
             List<String> ac = articleCategoryMapper.findByArticleId(article.getId());
 
             // 2.4添加到文章分类名称集合中
